requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
